requirements.txt
Django==5.0.4
psycopg2-binary==2.9.9
pandas==2.2.1
openpyxl==3.1.2
ortools==9.8.3296
xhtml2pdf==0.2.7
django-environ==0.11.2
